requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.22
alembic>=1.12.0
psycopg2-binary>=2.9.9

# LLM integration
langchain>=0.0.300
ollama>=0.1.5

# Utilities
loguru>=0.7.2
pyyaml>=6.0.1
tenacity>=8.2.3

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
httpx>=0.25.0

# Typing
mypy>=1.6.1
types-pyyaml>=6.0.12.12

# Linting
black>=23.10.1
flake8>=6.1.0
isort>=5.12.0

# Database
asyncpg>=0.28.0
